analytics/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
asyncio-nats-client==0.11.5
asyncpg==0.27.0
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.5.7
charset-normalizer==3.1.0
dataclasses-json==0.5.7
frozenlist==1.3.3
greenlet==2.0.2
idna==3.4
langchain==0.0.192
langchainplus-sdk==0.0.4
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.24.3
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.1
pandas==2.0.2
protobuf==3.20.3
pydantic==1.10.8
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==2.0.15
tenacity==8.2.2
tqdm==4.65.0
typing-extensions==4.6.3
typing-inspect==0.9.0
tzdata==2023.3
urllib3==2.0.3
websockets==11.0.3
yarl==1.9.2
